# Improve performance of TransactionLog.add()