--- a/hazelcast/src/main/java/com/hazelcast/transaction/impl/TransactionLog.java
+++ b/hazelcast/src/main/java/com/hazelcast/transaction/impl/TransactionLog.java
@@ -24,8 +24,7 @@ import com.hazelcast.spi.impl.operationservice.OperationService;
 import com.hazelcast.spi.impl.operationservice.impl.OperationServiceImpl;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.LinkedList;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.Map;
@@ -38,38 +37,28 @@ import java.util.concurrent.Future;
  * the TransactionLog will contains 4 {@link TransactionLogRecord} instances.
  *
  * planned optimization:
- * Most transaction will be small, but an linkedlist + hashmap is created. Instead use an array and do a
+ * Most transaction will be small, but an linkedhashmap is created. Instead use an array and do a
  * linear search in that array. When there are too many items added, then enable the hashmap.
  */
 public class TransactionLog {
 
-    private final List<TransactionLogRecord> recordList = new LinkedList<TransactionLogRecord>();
-    private final Map<Object, TransactionLogRecord> recordMap = new HashMap<Object, TransactionLogRecord>();
+    private final Map<Object, TransactionLogRecord> recordMap = new LinkedHashMap<>();
 
     public TransactionLog() {
     }
 
     public TransactionLog(List<TransactionLogRecord> transactionLog) {
-        //
-        recordList.addAll(transactionLog);
+        for (TransactionLogRecord record : transactionLog) {
+            add(record);
+        }
     }
 
     public void add(TransactionLogRecord record) {
         Object key = record.getKey();
-
-        // there should be just one tx log for the same key. so if there is older we are removing it
-        if (key != null) {
-            TransactionLogRecord removed = recordMap.remove(key);
-            if (removed != null) {
-                recordList.remove(removed);
-            }
+        if (key == null) {
+            key = new Object();
         }
-
-        recordList.add(record);
-
-        if (key != null) {
-            recordMap.put(key, record);
-        }
+        recordMap.put(key, record);
     }
 
     public TransactionLogRecord get(Object key) {
@@ -77,14 +66,11 @@ public class TransactionLog {
     }
 
     public List<TransactionLogRecord> getRecordList() {
-        return recordList;
+        return new ArrayList<>(recordMap.values());
     }
 
     public void remove(Object key) {
-        TransactionLogRecord removed = recordMap.remove(key);
-        if (removed != null) {
-            recordList.remove(removed);
-        }
+        recordMap.remove(key);
     }
 
     /**
@@ -93,12 +79,12 @@ public class TransactionLog {
      * @return the number of TransactionRecords.
      */
     public int size() {
-        return recordList.size();
+        return recordMap.size();
     }
 
     public List<Future> commit(NodeEngine nodeEngine) {
         List<Future> futures = new ArrayList<Future>(size());
-        for (TransactionLogRecord record : recordList) {
+        for (TransactionLogRecord record : recordMap.values()) {
             Future future = invoke(nodeEngine, record, record.newCommitOperation());
             futures.add(future);
         }
@@ -107,7 +93,7 @@ public class TransactionLog {
 
     public List<Future> prepare(NodeEngine nodeEngine) {
         List<Future> futures = new ArrayList<Future>(size());
-        for (TransactionLogRecord record : recordList) {
+        for (TransactionLogRecord record : recordMap.values()) {
             Future future = invoke(nodeEngine, record, record.newPrepareOperation());
             futures.add(future);
         }
@@ -116,7 +102,7 @@ public class TransactionLog {
 
     public List<Future> rollback(NodeEngine nodeEngine) {
         List<Future> futures = new ArrayList<Future>(size());
-        ListIterator<TransactionLogRecord> iterator = recordList.listIterator(size());
+        ListIterator<TransactionLogRecord> iterator = getRecordList().listIterator(size());
         while (iterator.hasPrevious()) {
             TransactionLogRecord record = iterator.previous();
             Future future = invoke(nodeEngine, record, record.newRollbackOperation());
@@ -135,13 +121,13 @@ public class TransactionLog {
     }
 
     public void commitAsync(NodeEngine nodeEngine, ExecutionCallback callback) {
-        for (TransactionLogRecord record : recordList) {
+        for (TransactionLogRecord record : recordMap.values()) {
             invokeAsync(nodeEngine, callback, record, record.newCommitOperation());
         }
     }
 
     public void rollbackAsync(NodeEngine nodeEngine, ExecutionCallback callback) {
-        for (TransactionLogRecord record : recordList) {
+        for (TransactionLogRecord record : recordMap.values()) {
             invokeAsync(nodeEngine, callback, record, record.newRollbackOperation());
         }
     }