Improve performance of `TransactionLog.add()` by avoiding `LinkedList.remove()` call.

Code may be considered cleaner/simpler now, but the reason for this change is performance. When in a single transaction entries were changed, and then changed again (`TransactionalMap.put()` called on the same entry), the `TransactionLog.add()` method did a `LinkedList.remove()` call which has complexity O(n). The new implementation that uses `LinkedHashMap` instead of `LinkedList`+`HashMap` is up to 20% faster on 1000-entry transaction, and up to three times faster on 10000-entry transaction. On small transactions performance is the same.

In [this gist](https://gist.github.com/gdela/48c235ccdd083944ae0e258547a81c1c), you can find benchmark that I used to make sure that this change improves performance. Here are the results:

Original Implementation:
```
MyTransactionBenchmark._10000_baseline  avgt    6   287,601 ±   5,284  ms/op
MyTransactionBenchmark._10000_backward  avgt    6   901,295 ±   9,973  ms/op
MyTransactionBenchmark._10000_forward   avgt    6   307,499 ±  20,451  ms/op
MyTransactionBenchmark._10000_random    avgt    6  1082,282 ± 164,960  ms/op
MyTransactionBenchmark._1000_baseline   avgt    6    28,056 ±   1,384  ms/op
MyTransactionBenchmark._1000_backward   avgt    6    35,975 ±   0,623  ms/op
MyTransactionBenchmark._1000_forward    avgt    6    30,154 ±   1,075  ms/op
MyTransactionBenchmark._1000_random     avgt    6    36,131 ±   1,472  ms/op
MyTransactionBenchmark._100_baseline    avgt    6     2,839 ±   0,078  ms/op
MyTransactionBenchmark._100_backward    avgt    6     2,984 ±   0,042  ms/op
MyTransactionBenchmark._100_forward     avgt    6     2,990 ±   0,155  ms/op
MyTransactionBenchmark._100_random      avgt    6     3,051 ±   0,112  ms/op
MyTransactionBenchmark._10_baseline     avgt    6      ,320 ±   0,010  ms/op
MyTransactionBenchmark._10_backward     avgt    6      ,348 ±   0,006  ms/op
MyTransactionBenchmark._10_forward      avgt    6      ,344 ±   0,011  ms/op
MyTransactionBenchmark._10_random       avgt    6      ,346 ±   0,010  ms/op
```

After Change:
```
MyTransactionBenchmark._10000_baseline  avgt    6  294,238 ± 13,064  ms/op
MyTransactionBenchmark._10000_backward  avgt    6  305,480 ± 15,631  ms/op
MyTransactionBenchmark._10000_forward   avgt    6  300,084 ±  5,390  ms/op
MyTransactionBenchmark._10000_random    avgt    6  301,633 ±  8,136  ms/op
MyTransactionBenchmark._1000_baseline   avgt    6   28,368 ±  1,159  ms/op
MyTransactionBenchmark._1000_backward   avgt    6   30,114 ±  1,072  ms/op
MyTransactionBenchmark._1000_forward    avgt    6   29,761 ±  0,653  ms/op
MyTransactionBenchmark._1000_random     avgt    6   29,659 ±  0,729  ms/op
MyTransactionBenchmark._100_baseline    avgt    6    2,802 ±  0,033  ms/op
MyTransactionBenchmark._100_backward    avgt    6    2,921 ±  0,097  ms/op
MyTransactionBenchmark._100_forward     avgt    6    2,927 ±  0,045  ms/op
MyTransactionBenchmark._100_random      avgt    6    3,059 ±  0,503  ms/op
MyTransactionBenchmark._10_baseline     avgt    6     ,317 ±  0,007  ms/op
MyTransactionBenchmark._10_backward     avgt    6     ,339 ±  0,004  ms/op
MyTransactionBenchmark._10_forward      avgt    6     ,344 ±  0,025  ms/op
MyTransactionBenchmark._10_random       avgt    6     ,332 ±  0,002  ms/op
```

As you can see, when there are 1000 or 10000 entries in a transaction, the performance after change is better and more stable.