@devOpsHazelcast 2019-05-30
Can one of the admins verify this patch?