@TomaszGaweda 2019-05-30
FYI, this patch is already battle-tested on many our environments. We didn't see any problems, etc.