@gdela 2019-07-10
Any chances of merging this rather simple pull request into Hazelcast 3.12.x? The performance difference is even more evident for transactions in which 100000 entries are being modified. And yes, we do have such large transaction in our project that uses Hazelcast.

Original implementation:
```
MyTransactionBenchmark._100000_baseline  avgt    6    3496,810 ±   569,101  ms/op
MyTransactionBenchmark._100000_backward  avgt    6  103513,763 ± 51709,247  ms/op
MyTransactionBenchmark._100000_forward   avgt    6    3665,922 ±   381,869  ms/op
MyTransactionBenchmark._100000_random    avgt    6  168188,393 ± 47363,600  ms/op
MyTransactionBenchmark._10000_random     avgt    6    1157,420 ±    42,528  ms/op
```

After Change:
```
MyTransactionBenchmark._100000_baseline  avgt    6  3383,798 ± 977,911  ms/op
MyTransactionBenchmark._100000_backward  avgt    6  3778,254 ± 412,208  ms/op
MyTransactionBenchmark._100000_forward   avgt    6  3718,097 ± 215,745  ms/op
MyTransactionBenchmark._100000_random    avgt    6  3727,629 ± 758,245  ms/op
```