@mmedenjak 2019-09-18
Merged, thank you for the PR again @gdela and the review @petrpleshachkov !